# internal/product/interfaces/http/product_handler.go
package http

import (
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
)

// ProductHandler handles product-related HTTP requests
type ProductHandler struct {
	productService interface{}
}

// NewProductHandler creates a new product handler
func NewProductHandler(productService interface{}) *ProductHandler {
	return &ProductHandler{
		productService: productService,
	}
}

// CreateProduct creates a new product
// @Summary Create a new product
// @Description Create a new product (Admin only)
// @Tags admin-products
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param product body application.CreateProductRequest true "Product data"
// @Success 201 {object} application.ProductResponse
// @Failure 400 {object} map[string]string
// @Failure 401 {object} map[string]string
// @Failure 403 {object} map[string]string
// @Router /admin/products [post]
func (h *ProductHandler) CreateProduct(c *gin.Context) {
	// This would handle product creation
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "Create product endpoint - to be implemented",
	})
}

// GetProduct retrieves a product by ID
// @Summary Get a product by ID
// @Description Get product details by ID (Public)
// @Tags products
// @Accept json
// @Produce json
// @Param id path int true "Product ID"
// @Success 200 {object} application.ProductResponse
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Router /products/{id} [get]
func (h *ProductHandler) GetProduct(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 32)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": "Invalid product ID",
		})
		return
	}

	// This would retrieve the product
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "Get product endpoint - to be implemented",
		"id":      id,
	})
}

// ListProducts retrieves all products with pagination
// @Summary List all products
// @Description Get a paginated list of all products (Public)
// @Tags products
// @Accept json
// @Produce json
// @Param offset query int false "Offset for pagination" default(0)
// @Param limit query int false "Limit for pagination" default(10)
// @Success 200 {object} application.ListProductsResponse
// @Router /products [get]
func (h *ProductHandler) ListProducts(c *gin.Context) {
	// This would handle product listing with pagination
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "List products endpoint - to be implemented",
	})
}

// UpdateProduct updates an existing product
// @Summary Update a product
// @Description Update an existing product (Admin only)
// @Tags admin-products
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "Product ID"
// @Param product body application.UpdateProductRequest true "Product data"
// @Success 200 {object} application.ProductResponse
// @Failure 400 {object} map[string]string
// @Failure 401 {object} map[string]string
// @Failure 403 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Router /admin/products/{id} [put]
func (h *ProductHandler) UpdateProduct(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 32)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": "Invalid product ID",
		})
		return
	}

	// This would update the product
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "Update product endpoint - to be implemented",
		"id":      id,
	})
}

// DeleteProduct deletes a product
// @Summary Delete a product
// @Description Delete a product (Admin only)
// @Tags admin-products
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "Product ID"
// @Success 200 {object} map[string]string
// @Failure 400 {object} map[string]string
// @Failure 401 {object} map[string]string
// @Failure 403 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Router /admin/products/{id} [delete]
func (h *ProductHandler) DeleteProduct(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 32)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": "Invalid product ID",
		})
		return
	}

	// This would delete the product
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "Delete product endpoint - to be implemented",
		"id":      id,
	})
}

// SearchProducts searches products
// @Summary Search products
// @Description Search products by name or other criteria (Public)
// @Tags products
// @Accept json
// @Produce json
// @Param query query string false "Search query"
// @Param category query string false "Category filter"
// @Param brand query string false "Brand filter"
// @Param min_price query number false "Minimum price"
// @Param max_price query number false "Maximum price"
// @Param offset query int false "Offset for pagination" default(0)
// @Param limit query int false "Limit for pagination" default(10)
// @Success 200 {object} application.ListProductsResponse
// @Router /products/search [get]
func (h *ProductHandler) SearchProducts(c *gin.Context) {
	// This would handle product search
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "Search products endpoint - to be implemented",
	})
}

// ListProductsByCategory retrieves products by category
func (h *ProductHandler) ListProductsByCategory(c *gin.Context) {
	category := c.Param("category")
	// This would handle product listing by category
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message":  "List products by category endpoint - to be implemented",
		"category": category,
	})
}

// UpdateStock updates product stock
func (h *ProductHandler) UpdateStock(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 32)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": "Invalid product ID",
		})
		return
	}

	// This would update product stock
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "Update stock endpoint - to be implemented",
		"id":      id,
	})
}

// ReduceStock reduces product stock
func (h *ProductHandler) ReduceStock(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 32)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": "Invalid product ID",
		})
		return
	}

	// This would reduce product stock
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "Reduce stock endpoint - to be implemented",
		"id":      id,
	})
}

// IncreaseStock increases product stock
func (h *ProductHandler) IncreaseStock(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 32)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": "Invalid product ID",
		})
		return
	}

	// This would increase product stock
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "Increase stock endpoint - to be implemented",
		"id":      id,
	})
}

// ActivateProduct activates a product
func (h *ProductHandler) ActivateProduct(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 32)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": "Invalid product ID",
		})
		return
	}

	// This would activate the product
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "Activate product endpoint - to be implemented",
		"id":      id,
	})
}

// DeactivateProduct deactivates a product
func (h *ProductHandler) DeactivateProduct(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 32)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": "Invalid product ID",
		})
		return
	}

	// This would deactivate the product
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "Deactivate product endpoint - to be implemented",
		"id":      id,
	})
}

// MarkAsFeatured marks a product as featured
func (h *ProductHandler) MarkAsFeatured(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 32)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": "Invalid product ID",
		})
		return
	}

	// This would mark product as featured
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "Mark as featured endpoint - to be implemented",
		"id":      id,
	})
}

// UnmarkAsFeatured removes featured status from a product
func (h *ProductHandler) UnmarkAsFeatured(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 32)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": "Invalid product ID",
		})
		return
	}

	// This would unmark product as featured
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "Unmark as featured endpoint - to be implemented",
		"id":      id,
	})
}

// IncrementViewCount increments product view count
func (h *ProductHandler) IncrementViewCount(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 32)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": "Invalid product ID",
		})
		return
	}

	// This would increment view count
	// For now, return a placeholder response
	c.JSON(http.StatusOK, gin.H{
		"message": "Increment view count endpoint - to be implemented",
		"id":      id,
	})
}